+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # tye-plugin Changelog
 
 ## [Unreleased]
+
+## [0.4.0]
 ### Added
 - Options to the `tye run` command: `--watch`, `--debug`, `--framework`
 - `tye build` and `tye push` actions to the Services tool window